LIBFCREPO-1268. Moved `plastron.files` package to the plastron-repo distribution package.
https://umd-dit.atlassian.net/browse/LIBFCREPO-1268
