+++ plastron-repo/tests/repo/test_pcdm.py
@@ -5,8 +5,7 @@ from uuid import uuid4
 import pytest
 
 from plastron.client import Client, Endpoint
-from plastron.files import StringSource
-from plastron.jobs import FileSpec, FileGroup
+from plastron.files import StringSource, FileSpec, FileGroup
 from plastron.repo import Repository, ResourceType
 from plastron.repo.pcdm import PCDMObjectResource
